chore: support scala-js and scala the same time
