+++ compile.sh
@@ -29,7 +29,7 @@ compile 'fpc' 'fpc -O3 fpc/code.pas'
 compile 'crystal' 'crystal build -o crystal/code --release crystal/code.cr'
 compile 'scala' 'scala-cli --power package scala/code.scala -f -o scala/code'
 compile 'scala' 'scala-cli --power package --native scala/code.scala -f -o scala/code-native --native-mode release-full'
-compile 'scala' 'scala-cli --power package --js scala/code.scala -f -o scala/code.js --js-mode fullLinkJS '
+compile 'scala' 'scala-cli --power package --js scala/codeJS.scala -f -o scala/code.js --js-mode fullLinkJS'
 compile 'scala' 'bun build --bytecode --compile scala/code.js --outfile scala/bun'
 compile 'ldc2' 'ldc2 -O3 -release -boundscheck=off -mcpu=native flto=thin d/code.d'
 compile 'odin' 'odin build odin/code.odin -o:speed -file -out:odin/code'
